cambios

+++ src/main/java/com/utp/DemoOratorIA/infraestructure/repositories/JPAAnalisisRepository.java
@@ -3,5 +3,5 @@ package com.utp.DemoOratorIA.infraestructure.repositories;
 import com.utp.DemoOratorIA.infraestructure.entities.AnalisisEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JPAAnalisisRepository extends JpaRepository<AnalisisEntity, Long> {
+public interface JPAAnalisisRepository extends JpaRepository<AnalisisEntity, Integer> {
 }
